Type mismatch corrected
git-svn-id: https://svn.iu.hio.no/projects/cfengine-2/trunk@603 0bacebff-10ef-0310-a86f-bc4055e020e4

+++ src/env_monitor.c
@@ -100,7 +100,7 @@ double LAST_DAY_ENTROPY = 0.0;
 struct Item *PREVIOUS_STATE = NULL;
 struct Item *ENTROPIES = NULL;
 
-short NO_FORK = false;
+int NO_FORK = false;
 double FORGETRATE = 0.6;
 
 int LASTQ[CF_OBSERVABLES];
